sstables: remove default params from sstable constructor
The goal is to construct sstables only via make_sstables
that will be moved to class sstables_manager in a later patch.

Defining the default values in both interfaces is unneeded
and may to lead to them going out of sync.

Therefore, have only make_sstables provide the default parameter values.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>