build: Fix the build on Ubuntu
The way the `pkg-config` executable works on Fedora and Ubuntu is
different, since on Fedora `pkg-config` is provided by the `pkgconf`
project.

In the build directory of Seastar, `seastar.pc` and `seastar-testing.pc`
are generated. `seastar` is a requirement of `seastar-testing`.

When pkg-config is invoked like this:

    pkg-config --libs build/release/seastar-testing.pc

the version of `pkg-config` on Fedora resolves the reference to
`seastar` in `Requires` to the `seastar.pc` in the same directory.

However, the version of `pkg-config` on Ubuntu 18.04 does not:

    Package seastar was not found in the pkg-config search path.
    Perhaps you should add the directory containing `seastar.pc'
    to the PKG_CONFIG_PATH environment variable
    Package 'seastar', required by '/seastar-testing', not found

To address the divergent behavior, we set the `PKG_CONFIG_PATH` variable
to point to the directory containing `seastar.pc`. With this change, I
was able to configure Scylla on both Fedora 29 and Ubuntu 18.04.

Fixes #4218

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <d7164bde2790708425ac6761154d517404818ecd.1550002959.git.jhaberku@scylladb.com>