configure: remove sanitize_libs and merge sanitize into opt
These are flags we want to pass to both compilation and linking. There
is nothing special about the fact that they are sanitizer related.

With {sanitize} being passed to the link, we don't need {sanitize_libs}.

We do need to make sure -fno-sanitize=vptr is the last one in the
command line. Before we were implicitly getting it from seastar, but
it is bad practice to get some sanitizer flags from seastar but not
others.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>