scylla-gdb.py: add scylla fiber command
The scylla fiber command traverses a continuation chain, given an
arbitrary task pointer.
Example (cropped for brevity):
(gdb) scylla fiber this
 #0  (task*) 0x0000600000550360 0x000000000468ac40 vtable for seastar...
 #1  (task*) 0x0000600000550300 0x00000000046c3778 vtable for seastar...
 #2  (task*) 0x00006000018af600 0x00000000046c37a0 vtable for seastar...
 #3  (task*) 0x00006000005502a0 0x00000000046c37f0 vtable for seastar...
 #4  (task*) 0x0000600001a65e10 0x00000000046c6b10 vtable for seastar...

scylla fiber can be passed any expression that evaluates to a task
pointer. C++ variables, raw adresses and GDB variables (e.g. $1) all
work.

The command works by scanning the task object for pointers. If a pointer
is found it is dereferenced. If successful it checks that the pointer
dereferences to a vtable, the class for which is a known task.
If this succeeds the found task is saved, the scan then recursively
proceeds to scan the newly found task until a task with no further
attached continuations is found.