tests: make sure that commitlog replay works after truncate.
Tomek and I recently had a discussion about whether or not a commitlog
replay would be safe after we dropped or truncated a table that is not
flushed (durable, but auto_snapshots being false).

While we agreed that would be the safe, we both agreed we would feel
better with a unit test covering that.

This patch adds such a test (btw, it passes)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190318223811.6862-1-glauber@scylladb.com>